PEP8 checking for new PRs

+++ doc/transform_methods/rbasex-pRn.py
@@ -16,7 +16,7 @@ ax.spines['right'].set_visible(False)
 plt.xlim((0, rmax))
 plt.xlabel('radius')
 
-r = np.linspace(0, rmax, rmax * 20 + 1)
+r = np.linspace(0, rmax, int(rmax * 20 + 1))
 r[0] = np.finfo(float).eps  # (to avoid division by 0)
 
 def sqrtp(x):
@@ -61,5 +61,5 @@ plt.ylim(bottom=0)
 plt.legend()
 plt.tight_layout()
 
-#plt.savefig('rbasex-pRn.svg')
-#plt.show()
+# plt.savefig('rbasex-pRn.svg')
+# plt.show()
